Handle BA failures in phased action runner
Move exception handling in case of failures in BuildActions from
ProviderConnection to ClientProvidedPhasedActionRunner.
