+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/ProviderConnection.java
@@ -16,7 +16,6 @@
 
 package org.gradle.tooling.internal.provider;
 
-import org.gradle.api.Transformer;
 import org.gradle.api.internal.StartParameterInternal;
 import org.gradle.api.logging.LogLevel;
 import org.gradle.initialization.BuildCancellationToken;
@@ -46,9 +45,6 @@ import org.gradle.tooling.internal.build.DefaultBuildEnvironment;
 import org.gradle.tooling.internal.consumer.parameters.FailsafeBuildProgressListenerAdapter;
 import org.gradle.tooling.internal.consumer.versioning.ModelMapping;
 import org.gradle.tooling.internal.gradle.DefaultBuildIdentifier;
-import org.gradle.tooling.internal.protocol.BuildExceptionVersion1;
-import org.gradle.tooling.internal.protocol.InternalBuildActionFailureException;
-import org.gradle.tooling.internal.protocol.InternalBuildCancelledException;
 import org.gradle.tooling.internal.protocol.PhasedActionResultListener;
 import org.gradle.tooling.internal.protocol.InternalBuildAction;
 import org.gradle.tooling.internal.protocol.InternalBuildActionVersion2;
@@ -159,11 +155,6 @@ public class ProviderConnection {
         try {
             return run(action, cancellationToken, listenerConfig, new PhasedActionEventConsumer(failsafePhasedActionResultListener, payloadSerializer, listenerConfig.buildEventConsumer),
                 providerParameters, params);
-        } catch (BuildExceptionVersion1 e) {
-            // Failures in BuildActions in a PhasedAction will be wrapped inside a BuildExceptionVersion1 since the build itself will fail.
-            // This adapter unwraps the expected exception (thrown by ClientProvidedPhasedActionRunner) so it is in the correct format defined by InternalPhasedActionConnection#run
-            // TODO: instead of unwrap this exception here, introduce API in org.gradle.internal.invocation.BuildController to catch these partial exceptions and set as a build result failure
-            throw new PhasedActionExceptionTransformer().transform(e);
         } finally {
             failsafePhasedActionResultListener.rethrowErrors();
         }
@@ -325,17 +316,6 @@ public class ProviderConnection {
                     delegate.dispatch(message);
                 }
             }
-        }
-    }
-
-    class PhasedActionExceptionTransformer implements Transformer<RuntimeException, RuntimeException> {
-        public RuntimeException transform(RuntimeException e) {
-            for (Throwable t = e; t != null; t = t.getCause()) {
-                if (t instanceof InternalBuildActionFailureException || t instanceof InternalBuildCancelledException) {
-                    return (RuntimeException) t;
-                }
-            }
-            return e;
         }
     }
 }
