+++ subprojects/tooling-api-builders/src/test/groovy/org/gradle/tooling/internal/provider/runner/ClientProvidedPhasedActionRunnerTest.groovy
@@ -137,9 +137,15 @@ class ClientProvidedPhasedActionRunnerTest extends Specification {
 
         then:
         1 * projectsLoadedAction.execute(_) >> { throw new RuntimeException() }
-        thrown(InternalBuildActionFailureException)
         0 * projectsEvaluatedAction.execute(_)
         0 * buildFinishedAction.execute(_)
+        1 * buildController.setResult(_) >> { args ->
+            def it = args[0]
+            assert it instanceof BuildActionResult
+            assert it.failure == serializedFailure
+            assert it.result == null
+            buildController.hasResult() >> true
+        }
         0 * buildEventConsumer.dispatch({
             it instanceof PhasedBuildActionResult &&
                 it.phase == PhasedActionResult.Phase.PROJECTS_LOADED
@@ -160,14 +166,14 @@ class ClientProvidedPhasedActionRunnerTest extends Specification {
 
         then:
         1 * projectsLoadedAction.execute(_) >> { throw new RuntimeException() }
-        thrown(InternalBuildActionFailureException)
+        1 * payloadSerializer.serialize({ it instanceof InternalBuildActionFailureException })
 
         when:
         runner.run(clientProvidedPhasedAction, buildController)
 
         then:
         1 * projectsLoadedAction.execute(_) >> { throw new BuildCancelledException() }
-        thrown(InternalBuildCancelledException)
+        1 * payloadSerializer.serialize({ it instanceof InternalBuildCancelledException })
     }
 
     def "action not run if null"() {
